Tests: Super admin can create and view supervisors and sales reps

+++ main/app/Modules/SuperAdmin/Traits/IsAStaff.php
@@ -10,6 +10,8 @@ use Illuminate\Support\Facades\Route;
 use Illuminate\Database\Eloquent\Builder;
 use App\Modules\SuperAdmin\Models\ActivityLog;
 use App\Modules\SuperAdmin\Transformers\StaffTransformer;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Modules\SuperAdmin\Http\Requests\CreateStaffRequest;
 
 /**
  * A trait to make a model commentable
@@ -17,14 +19,16 @@ use App\Modules\SuperAdmin\Transformers\StaffTransformer;
 trait IsAStaff
 {
 
+  use AuthorizesRequests;
+
   static function findByEmail(string $email)
   {
     return self::whereEmail($email)->first();
   }
 
-  public function is_verified()
+  public function is_verified(): bool
   {
-    return $this->verified_at !== null;
+    return $this->verified_at !== null || $this->isSuperAdmin();
   }
 
   /**
@@ -34,58 +38,44 @@ trait IsAStaff
    */
   static function staffRoutes()
   {
-    $userType = Str::of(class_basename(self::class))->snake()->plural();
-    Route::name('superadmin.manage_staff.')->prefix($userType->slug())->group(function () use ($userType) {
-      Route::get('', [self::class, 'getAllStaff'])->name($userType)->defaults('ex', __e('ss', 'aperture'));
-      Route::post('create', [self::class, 'createStaff'])->name($userType . '.create');
-      Route::put('{staff}/edit', [self::class, 'editStaff'])->name($userType . '.edit');
-      Route::put('{staff}/suspend', [self::class, 'suspendStaff'])->name($userType . '.suspend');
-      Route::put('{staff}/restore', [self::class, 'restoreStaff'])->name($userType . '.reactivate');
-      Route::delete('{staff}/delete', [self::class, 'deleteStaff'])->name($userType . '.delete');
+    Route::middleware('auth:super_admin')->group(function () {
+      Route::get('list', [self::class, 'getAllStaff'])->name('list')->defaults('ex', __e('ss', 'aperture'));
+      Route::post('create', [self::class, 'createStaff'])->name('create');
+      Route::put('{staff}/edit', [self::class, 'editStaff'])->name('edit');
+      Route::put('{staff}/suspend', [self::class, 'suspendStaff'])->name('suspend');
+      Route::put('{staff}/activate', [self::class, 'restoreStaff'])->name('activate');
+      Route::delete('{staff}/delete', [self::class, 'deleteStaff'])->name('delete');
+      Route::put('{staff}/restore', [self::class, 'restoreStaff'])->name('restore');
     });
   }
 
   public function getAllStaff(Request $request)
   {
+    $this->authorize('viewAny', self::class);
+
     return Inertia::render('SuperAdmin,ManageStaff/Manage' . Str::of(__CLASS__)->afterLast('\\')->plural(), [
-      // (string)Str::of(class_basename(self::class))->snake()->plural() => (new StaffTransformer)->collectionTransformer(self::all(), 'transformForSuperAdminViewSalesReps'),
+      (string)Str::of(class_basename(self::class))->snake()->plural() => (new StaffTransformer)->collectionTransformer(self::all(), 'transformForSuperAdminViewSalesReps'),
     ]);
   }
 
-  public function createStaff(Request $request)
+  public function createStaff(CreateStaffRequest $request)
   {
-    $userType = Str::of(class_basename(self::class))->snake()->plural();
+    $this->authorize('create', self::class);
 
-    $validated = $request->validate([
-      'full_name' => 'required|string|max:20',
-      'email' => 'required|email|max:50|unique:' . $userType . ',email',
-      'avatar' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg',
-      'password' => ''
-    ]);
-
-    $validated['password'] = $userType->slug();
-
-    if ($request->hasFile('avatar')) {
-      $validated['avatar'] = compress_image_upload('avatar', 'user-avatars/', 'user-avatars/thumbs/', 1400, true, 50)['img_url'];
-    }
+    $userType =  Str::of(class_basename(self::class))->snake()->plural();
 
     try {
-
-      $staff = self::create($validated);
-
-      ActivityLog::notifySuperAdmins($request->user()->full_name . ' created a ' . $userType->singular()->slug(' ') . ' account for ' . $staff->full_name);
-
-      return back()->withFlash(['success' => $userType->singular()->slug(' ') . ' account created']);
-    } catch (Throwable $e) {
-      if (app()->environment() == 'local') {
-        return back()->withFlash(['error' => $e->getMessage()]);
-      }
-      return back()->withFlash(['error' => 'Error occurred']);
+      $request->createStaff(self::class);
+    } catch (\Throwable $th) {
+      return back()->withFlash(['error' => $th->getMessage()]);
     }
+
+    return redirect()->route($userType->singular()->slug('') .'.list')->withFlash(['success' => $userType->singular()->replaceFirst('_',' ')->title() . ' account created. Activate the account so the user can login']);
   }
 
   public function editStaff(Request $request, self $staff)
   {
+    $this->authorize('update', $staff);
 
     $userType = Str::of(class_basename(self::class))->snake()->plural();
 
@@ -116,6 +106,20 @@ trait IsAStaff
 
   public function suspendStaff(self $staff)
   {
+    $this->authorize('suspend', $staff);
+
+    ActivityLog::logUserActivity(auth()->user()->email . ' suspended the account of ' . $staff->email);
+
+    $staff->is_active = false;
+    $staff->save();
+
+    return back()->withFlash(['success' => 'User account suspended']);
+  }
+
+  public function activateStaff(self $staff)
+  {
+    $this->authorize('activate', $staff);
+
     ActivityLog::logUserActivity(auth()->user()->email . ' suspended the account of ' . $staff->email);
 
     $staff->is_active = false;
@@ -126,6 +130,8 @@ trait IsAStaff
 
   public function restoreStaff(self $staff)
   {
+    $this->authorize('restore', $staff);
+
     $staff->is_active = true;
     $staff->save();
 
@@ -136,6 +142,8 @@ trait IsAStaff
 
   public function deleteStaff(self $staff)
   {
+    $this->authorize('delete', $staff);
+
     ActivityLog::logUserActivity(auth()->user()->email . ' permanently deleted the account of ' . $staff->email);
 
     $staff->forceDelete();
